small filelogger bugfix
git-svn-id: http://svn.3splooges.com/romraider-arch/trunk@297 d2e2e1cd-ba16-0410-be16-b7c4453c7c2d

+++ src/enginuity/logger/ui/handler/FileUpdateHandler.java
@@ -33,8 +33,8 @@ public final class FileUpdateHandler implements DataUpdateHandler, ConvertorUpda
 
     public void handleDataUpdate(EcuData ecuData, byte[] bytes, long timestamp) {
         EcuDataConvertor convertor = ecuData.getSelectedConvertor();
-        int value = (int) convertor.convert(bytes);
-        checkStartStopFileLogging(ecuData, value);
+        double value = convertor.convert(bytes);
+        checkStartStopFileLogging(ecuData, (int) value);
         if (fileLogger.isStarted()) {
             currentLine.updateParamValue(ecuData, convertor.format(value), timestamp);
             if (currentLine.isFull()) {
